backend/requirements.txt
crewai>=0.130.0
python-dotenv>=1.0.1
openai>=1.0.0
litellm>=1.0.0
apscheduler>=3.10.4
fastapi>=0.116.0
uvicorn>=0.30.0
email-validator>=2.3.0
fastapi-sso>=0.20.0
langchain-community>=0.4.0
langchain-text-splitters>=1.1.0
langchain-mongodb>=0.11.0
langchain-huggingface>=1.2.0
pymongo>=4.16.0
sentence-transformers>=5.2.0
pypdf>=6.7.0
